backend/requirements.txt
Authlib==1.6.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.2
marshmallow==4.0.0
psutil==7.0.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pywin32==311
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WMI==1.4.9
wsproto==1.2.0
